Save a Fact record to the database
Given a Fact record of the form (Fact. name value timestamp certname)
saves it to the database when the (save! ...) function is called on
the record.

This will provide the abstraction required for splitting out the
'command' layer and the 'query' layer, but still providing a
nicer (interfaced) data structure to play with on either side.